Added menu tab to navigate on How It Works page
